+++ packages/thumbprint-react/index.ts
@@ -2,21 +2,19 @@ export { UserAvatar, EntityAvatar } from './components/Avatar/index';
 export { default as Button, TextButton } from './components/Button/index';
 export { BannerAlert, InPageAlert } from './components/Alert/index';
 export { default as ButtonRow } from './components/ButtonRow/index';
+export { default as DatePicker, Calendar } from './components/Calendar/index';
 export { default as Carousel } from './components/Carousel/index';
 export { default as Checkbox } from './components/Checkbox/index';
-export { default as DatePicker } from './components/DatePicker/index';
+export { default as Select, Dropdown } from './components/Dropdown/index';
 export { default as FormNote } from './components/FormNote/index';
 export { default as HorizontalRule } from './components/HorizontalRule/index.jsx';
 export { Grid, GridColumn } from './components/Grid/index';
 export { default as Image } from './components/Image/index';
-export { default as Input, InputIcon, InputClearButton } from './components/Input/index';
 export { default as InputRow, InputRowContext } from './components/InputRow/index.jsx';
 export { default as Label } from './components/Label/index';
 export { default as Link, ThemedLink } from './components/Link/index.jsx';
 export { List, ListItem } from './components/List/index';
 export { default as LoaderDots } from './components/LoaderDots/index';
-export { default as ModalBase } from './components/ModalBase/index.jsx';
-export { default as ModalCurtain } from './components/ModalCurtain/index.jsx';
 export {
     default as ModalDefault,
     ModalDefaultHeader,
@@ -26,7 +24,17 @@ export {
     ModalDefaultContentFullBleed,
     ModalDefaultFooter,
     ModalDefaultAnimatedWrapper,
-} from './components/ModalDefault/index.jsx';
+    Modal,
+    ModalHeader,
+    ModalTitle,
+    ModalDescription,
+    ModalContent,
+    ModalContentFullBleed,
+    ModalFooter,
+    ModalAnimatedWrapper,
+} from './components/Modal/index.jsx';
+export { default as ModalBase } from './components/ModalBase/index.jsx';
+export { default as ModalCurtain } from './components/ModalCurtain/index.jsx';
 export { default as ModalStandard } from './components/ModalStandard/index.jsx';
 export {
     default as Popover,
@@ -37,7 +45,6 @@ export {
 } from './components/Popover/index';
 export { default as Pill } from './components/Pill/index.jsx';
 export { default as Radio } from './components/Radio/index';
-export { default as Select } from './components/Select/index';
 export {
     default as ServiceCard,
     ServiceCardDescription,
@@ -45,7 +52,15 @@ export {
     ServiceCardImage,
 } from './components/ServiceCard/index';
 export { default as StarRating } from './components/StarRating/index';
-export { default as Textarea } from './components/Textarea/index';
+export { default as Textarea, TextArea } from './components/TextArea/index';
+export {
+    default as Input,
+    InputIcon,
+    InputClearButton,
+    TextInput,
+    TextInputIcon,
+    TextInputClearButton,
+} from './components/TextInput/index';
 export { default as Tooltip } from './components/Tooltip/index';
 export { Title, Text } from './components/Type/index';
 export { default as Wrap } from './components/Wrap/index';